+++ src/mdsvex.svelte
@@ -1,7 +1,6 @@
 <script context="module">
     import Img from '$lib/components/Img.svelte'
     import A from '$lib/components/A.svelte'
-
     export { Img as img }
     export { A as a }
 </script>